+++ cpp/open3d/pipelines/color_map/ColorMapUtils.h
@@ -44,7 +44,7 @@ namespace color_map {
 std::vector<std::shared_ptr<geometry::Image>> CreateDepthBoundaryMasks(
         const std::vector<std::shared_ptr<geometry::Image>>& images_depth,
         double depth_threshold_for_discontinuity_check,
-        double half_dilation_kernel_size_for_discontinuity_map);
+        int half_dilation_kernel_size_for_discontinuity_map);
 
 std::tuple<std::vector<std::vector<int>>, std::vector<std::vector<int>>>
 CreateVertexAndImageVisibility(
